More integration test cases
