+++ src/test/resources/pattern-match/input.scala
@@ -1,0 +1,13 @@
+package foo
+
+object Foo {
+
+  def bar(x: Int): String = x match {
+    case 0 => "zero"
+    case 1 =>
+      val dunno = "one?"
+      ???
+    case _ => ???
+  }
+
+}
