+++ src/test/resources/polymorphic/input.scala
@@ -1,0 +1,10 @@
+package foo
+
+object Foo {
+
+  def polymorphic[A](x: A, y: A): String = {
+    if (x == y) "equal"
+    else ???
+  }
+
+}
